+++ src/test/java/com/strangegrotto/wealthdraft/TestFileProvider.java
@@ -1,4 +1,4 @@
-package com.strangegrotto.wealthdraft.networth.projections;
+package com.strangegrotto.wealthdraft;
 
 import java.net.URL;