+++ src/test/java/com/strangegrotto/wealthdraft/networth/projections/ExpectedExampleProjections.java
@@ -1,21 +1,17 @@
 package com.strangegrotto.wealthdraft.networth.projections;
 
-import com.strangegrotto.wealthdraft.networth.BankAccountAssetChange;
-import com.strangegrotto.wealthdraft.networth.ExpectedAssetsWithHistoryInfo;
-import com.strangegrotto.wealthdraft.networth.ImmBankAccountAssetChange;
+import com.strangegrotto.wealthdraft.networth.assets.ExpectedExampleAssetsWithHistory;
 import com.strangegrotto.wealthdraft.networth.ImmBankAccountAssetChange;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Map;
-import java.util.Optional;
-import java.util.SortedMap;
 import java.util.TreeMap;
 
 /**
  * Encapsulates what data we expect the example projections file to have
  */
-public class ExpectedProjectionsInfo {
+public class ExpectedExampleProjections {
     public static final String SELL_ALL_BTC_3Y_ID = "btc-3y";
     public static final String SELL_HALF_BTC_1Y_ID = "half-btc-1y";
     public static final String SELL_OTHER_HALF_BTC_2Y_ID = "half-btc-2y";
@@ -42,8 +38,8 @@ public class ExpectedProjectionsInfo {
                 .build();
         var allBtc3yAllChanges = new TreeMap<LocalDate, Map<String, AssetChange>>(Map.of(
                 today.plusYears(3), Map.of(
-                        ExpectedAssetsWithHistoryInfo.BITCOIN_ACCOUNT_ASSET_ID, allBtc3yBtcChange,
-                        ExpectedAssetsWithHistoryInfo.BANK_ACCOUNT_ASSET_ID, allBtc3yBankChange
+                        ExpectedExampleAssetsWithHistory.BITCOIN_HOLDING_ID, allBtc3yBtcChange,
+                        ExpectedExampleAssetsWithHistory.BANK_ACCOUNT_ID, allBtc3yBankChange
                 )
         ));
         SELL_ALL_BTC_3Y_SCENARIO = ImmProjectionScenario.of(
@@ -64,8 +60,8 @@ public class ExpectedProjectionsInfo {
                 .build();
         var halfBtc1yAllChanges = new TreeMap<LocalDate, Map<String, AssetChange>>(Map.of(
                 today.plusYears(1), Map.of(
-                        ExpectedAssetsWithHistoryInfo.BITCOIN_ACCOUNT_ASSET_ID, halfBtc1yBtcChange,
-                        ExpectedAssetsWithHistoryInfo.BANK_ACCOUNT_ASSET_ID, halfBtc1yBankChange
+                        ExpectedExampleAssetsWithHistory.BITCOIN_HOLDING_ID, halfBtc1yBtcChange,
+                        ExpectedExampleAssetsWithHistory.BANK_ACCOUNT_ID, halfBtc1yBankChange
                 )
         ));
         SELL_HALF_BTC_1Y_SCENARIO = ImmProjectionScenario.of(
@@ -86,8 +82,8 @@ public class ExpectedProjectionsInfo {
                 .build();
         var otherHalfBtc2yAllChanges = new TreeMap<LocalDate, Map<String, AssetChange>>(Map.of(
                 today.plusYears(2), Map.of(
-                        ExpectedAssetsWithHistoryInfo.BITCOIN_ACCOUNT_ASSET_ID, otherHalfBtc2yBtcChange,
-                        ExpectedAssetsWithHistoryInfo.BANK_ACCOUNT_ASSET_ID, otherHalfBtc2yBankChange
+                        ExpectedExampleAssetsWithHistory.BITCOIN_HOLDING_ID, otherHalfBtc2yBtcChange,
+                        ExpectedExampleAssetsWithHistory.BANK_ACCOUNT_ID, otherHalfBtc2yBankChange
                 )
         ));
         otherHalfBtc2yAllChanges.putAll(halfBtc1yAllChanges);