+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Update to Java 11 to get the lovely `var` keyword
 * Switch to using BigDecimal for all networth calculations
 * Add custom `Immutables` style, and switch all networth calculation code over to it
+* Add a deserializer with tests to `AssetsWithHistory`, so arbitrary asset history can be added
 
 ## 0.5.0
 * Allow users to base scenarios on other scenarios